Fixed Cef crash, finished implementing the new spotify api

+++ OpenLyricsClient/Backend/Structure/SpotifyAccess.cs
@@ -6,7 +6,7 @@ namespace OpenLyricsClient.Backend.Structure
     {
         public string AccessToken { get; set; }
         public string RefreshToken { get; set; }
-        public int SpotifyExpireTime { get; set; }
+        public long SpotifyExpireTime { get; set; }
         public bool IsSpotifyConnected { get; set; }
     }
 }
